# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiErrorResponse {
  error: string;
}

export interface OddsOutcome {
  name: string;
  price: number;
  /** @nullable */
  point?: number | null;
}

export interface OddsMarket {
  key: string;
  outcomes: OddsOutcome[];
}

export interface OddsGame {
  id: string;
  sport: string;
  homeTeam: string;
  awayTeam: string;
  commenceTime: string;
  markets: OddsMarket[];
}

export interface EspnGame {
  id: string;
  sport: string;
  name: string;
  shortName: string;
  status: string;
  startsAt: string;
  /** @nullable */
  homeTeam?: string | null;
  /** @nullable */
  awayTeam?: string | null;
  /** @nullable */
  homeScore?: number | null;
  /** @nullable */
  awayScore?: number | null;
  /** @nullable */
  homeTeamId?: string | null;
  /** @nullable */
  awayTeamId?: string | null;
  /** @nullable */
  venue?: string | null;
}

export interface InjuryEntry {
  player: string;
  /** @nullable */
  position?: string | null;
  status: string;
  description: string;
}

export interface InjuryReport {
  team: string;
  teamAbbr: string;
  entries: InjuryEntry[];
}

export interface WeatherSnapshot {
  tempF: number;
  condition: string;
  windMph: number;
  humidity: number;
  precipChance: number;
  /** @nullable */
  city?: string | null;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  system: 'system',
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

/**
 * Optional structured context (selected sports, current parlay legs, etc.)
 */
export type ChatInputContext = { [key: string]: unknown };

export interface ChatInput {
  messages: ChatMessage[];
  /** Optional structured context (selected sports, current parlay legs, etc.) */
  context?: ChatInputContext;
}

export type GetOddsParams = {
/**
 * Sport id (nfl, nba, mlb, nhl, soccer, ncaaf, ncaab, ufc)
 */
sport: string;
};

export type GetGamesParams = {
sport: string;
};

export type GetInjuriesParams = {
sport: string;
};

export type GetWeatherParams = {
lat: number;
lon: number;
};