# lib/api-zod/src/generated/types/chatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatInputContext } from './chatInputContext';
import type { ChatMessage } from './chatMessage';

export interface ChatInput {
  messages: ChatMessage[];
  /** Optional structured context (selected sports, current parlay legs, etc.) */
  context?: ChatInputContext;
}