# lib/api-zod/src/generated/types/espnGame.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EspnGame {
  id: string;
  sport: string;
  name: string;
  shortName: string;
  status: string;
  startsAt: string;
  /** @nullable */
  homeTeam?: string | null;
  /** @nullable */
  awayTeam?: string | null;
  /** @nullable */
  homeScore?: number | null;
  /** @nullable */
  awayScore?: number | null;
  /** @nullable */
  homeTeamId?: string | null;
  /** @nullable */
  awayTeamId?: string | null;
  /** @nullable */
  venue?: string | null;
}