# lib/api-zod/src/generated/types/injuryEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InjuryEntry {
  player: string;
  /** @nullable */
  position?: string | null;
  status: string;
  description: string;
}