# lib/api-zod/src/generated/types/injuryReport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InjuryEntry } from './injuryEntry';

export interface InjuryReport {
  team: string;
  teamAbbr: string;
  entries: InjuryEntry[];
}