# lib/api-zod/src/generated/types/oddsMarket.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OddsOutcome } from './oddsOutcome';

export interface OddsMarket {
  key: string;
  outcomes: OddsOutcome[];
}