# lib/api-zod/src/generated/types/oddsOutcome.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OddsOutcome {
  name: string;
  price: number;
  /** @nullable */
  point?: number | null;
}