# lib/api-zod/src/generated/types/weatherSnapshot.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WeatherSnapshot {
  tempF: number;
  condition: string;
  windMph: number;
  humidity: number;
  precipChance: number;
  /** @nullable */
  city?: string | null;
}